Extract path related types from types/utils.ts
