+++ src/types/index.ts
@@ -4,6 +4,7 @@ export * from './events';
 export * from './fieldArray';
 export * from './fields';
 export * from './form';
+export * from './path';
 export * from './resolvers';
 export * from './utils';
 export * from './validator';
