+++ src/types/validator.ts
@@ -1,6 +1,6 @@
 import { Message } from './errors';
 import { FieldValues, InternalFieldName } from './fields';
-import { FieldPath, FieldPathValue } from './utils';
+import { FieldPath, FieldPathValue } from './path';
 
 export type ValidationValue = boolean | number | string | RegExp;
